Updated PageRank.scala
Implemented pagerank using the MonteCarlo method described in Appendinx B. Implemented in MonteCarlo.scala, called in PageRank.scala. `pagerank` is a curried function! The number of users walking through Monte Carlo can be changed with the first parameter: `pagerank(10)`  or it will default to 10,000: `pagerank()`. This will return the true pagerank function, pass `pages` to it: `pagerank()(pages)`